requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.5
python-dotenv==1.0.0
plivo==4.57.0
python-engineio==4.8.0
python-socketio==5.10.0
lxml==4.9.3
